+++ tbot223_base/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.0rc1"
+__version__ = "1.0.0rc2"
 
 from tbot223_base.exception_tracker import ExceptionTracker, ExceptionTrackerDecorator
 from tbot223_base.result import Result, ResultStatus, ResultUnwrapException